// apps/web/lib/db/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.4"
  }
  public: {
    Tables: {
// ...
      sources: {
        Row: {
          created_at: string
          fetch_method: string
          id: string
          is_active: boolean
          jurisdiction: string
          name: string
          scrape_config: Json | null
          tier: string
          url: string
        }
        Insert: {
          created_at?: string
          fetch_method: string
          id?: string
          is_active?: boolean
          jurisdiction?: string
          name: string
          scrape_config?: Json | null
          tier: string
          url: string
        }
        Update: {
          created_at?: string
          fetch_method?: string
          id?: string
          is_active?: boolean
          jurisdiction?: string
          name?: string
          scrape_config?: Json | null
          tier?: string
          url?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}